backend: rename `kv` package to `backend` package
